backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]==0.21.0
python-multipart==0.0.6
firebase-admin==6.4.0
google-generativeai==0.3.2
pydantic>=2.0.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.24.0
pytest==7.4.4
pytest-asyncio==0.23.2
typing-extensions==4.9.0
# Additional dependencies from sih integration
Flask==2.3.3
Flask-CORS==4.0.0